saga_asset_foundry/requirements.txt
fastapi==0.110.0
uvicorn==0.27.1
pillow==10.2.0
python-multipart==0.0.9
httpx==0.27.0
